Refactor: separate handle rand listener
